Add response ID parsing tests
